lab4 mistakes fixed

+++ src/print_helper.h
@@ -3,6 +3,7 @@
 #define PRINT_HELPER
 
 int print_ascii_tbl (FILE *stream);
-int print_for_human (FILE *stream, const unsigned char *array, const int len);
+//int print_for_human (FILE *stream, const unsigned char *array, const int len);
+int print_for_human (FILE *stream, const unsigned char *array, const size_t len);
 
 #endif /* PRINT_HELPER */
